Use manta-style args and environs
Also, use dashdash for parsing args (just to make it a little easier to
get the manta args in there).

Perhaps consider moving the first argument to a `-f` to match `mput`
syntax?  I personally like it as two positional args, but it is less
consistent with the "official" manta toolkit.